move loginPage to directory in Pages
