+++ src/components/Root.tsx
@@ -2,7 +2,7 @@ import React, { useContext, useEffect } from "react";
 import { Header } from "./Header";
 import { useNavigate, Outlet } from "react-router-dom";
 import { AuthContext } from "../contexts/AuthProvider";
-import { LoginPage } from "../pages/LoginPage";
+import { LoginPage } from "../pages/login/LoginPage";
 import { WelcomePage } from "../pages/WelcomePage";
 
 export default function Root() {
